Add output plugin for InfluxDB

+++ common/common.go
@@ -178,11 +178,11 @@ func Trim(value *string) string {
 	return *value
 }
 
-func Percentage(value uint64, max uint64) int {
+func Percentage(value int64, max uint64) float64 {
 	v := float64(value)
 	m := float64(max)
 	if v >= 0 && m > 0 {
-		return int((v / m) * 100)
+		return (v / m) * 100
 	}
 	return 0
 }
